[feat/#25] refactor: 더미 데이터 생성 유틸 수정
- DummyMemberGenerator: AtomicInteger 타입으로 변경
- DummyStayGenerator: 패키지 이동, 리팩토링
